// src/interfaces/int.ts
export interface WardsApplicant {
  wlastName: string
  wfirstName: string
  wotherName: string
  wdob: Date | null
  wgender: string
  wcity_ob: string
  wcountry_ob: string
  index: number | null
}
// ...
export interface Applicant {
  created_at?: Date
  apl_id?: string
  plastName: string
  pfirstName: string
  potherName: string
  pdob: Date | null
  pcity_ob: string
  pcountry_ob: string
  pgender: string
  pconf_code: string
  pemail: string
  ppassport_number: string
  passport_ex: Date | null
  pcountry_live_today: string
  peducation_level: string
  ppostal: string
  pmarital_status: string
  children_number: number
  fullName: string
  user_id: string
  pcontact: string
  wards: WardsApplicant[]
  totalPayment: number
  pother_contact: string
  psocial_media: {
    facebook: string
    instagram: string
    twitter: string
  }
  aplImg_path: {
    primePath: string[]
    secPath: string[]
    wardsPath: string[]
  }
  slastName: string
  sfirstName: string
  sotherName: string
  scity_ob: string
  scountry_ob: string
  scontact: string
  sgender: string
  sdob: Date | null
  location?: string
}